Add test case requiring subtracting "arm".
This contains three images: background, an image with an uncompressed
slinky, and an image with a compressed slinky.

Note that fg2.jpg contains two black objects to represent "arms" that
compress the object. Thus the object detector will need to ignore those
areas to correctly identify the resulting compressed object.