Change names and add rectangle compare method.
The class was changed from ObjectDetector to SegmentedObject, which
should make it clearer to the user to use separate instances for
different image pairs. Indeed, the set_image method was removed to
prevent recycling the same SegmentedObject.

The module name was changed to obj_detect to move in line with
convention.

The rectangle compare method is designed help compare the sizes of
different SegmentedObjects.