Adjust rectangle dimension method.
Previously, the axis upon which the width/height dimensions that the min
area rectangle would be calculated could be arbitrary, which made it
difficult to figure out whether the axes were "aligned" between the
uncompressed object and compressed object images. 

The solution was to make the width represent the axis "closest" (by
angle) to the x-axis, and the height the axis "closest" to the y-axis.
This should ensure that axes remain "constant" as the object gets
compressed, save when it is at a 45 degree angle with the xy axis.